testing/20150629.html

<!DOCTYPE html>
<title>IRC logs: w3c / #testing / 20150629</title>
<meta name="viewport" content="width=device-width">
<style>
ol { font-family: monospace; color: #000; font-size: 90%; list-style: none; margin: 0; padding: 0; }
li { margin: 0; padding: 0; white-space: pre-wrap; word-wrap: break-word; }
li:target { background: #f90; }
li.hide { display: none; }
li.flagged { background: yellow; }
li.filtered { color: #aaa; font-size: 60%; }
li.filtered:hover { font-size: 100%; color: #000; height: auto; }
li span { display: inline; margin: 0; padding: 0 10px; background: #eee; color: #000; text-decoration: none; cursor: pointer; }
li.flagged span { background: #f00; }
li:hover span { background: #ff0 }
input#filter { margin: 0 10px; }
label { cursor: pointer; }
p span { background: #ff0; margin-left: 10px; }
</style>
<h1 id="w3c/testing.20150629.log"><a href="/irc-logs/">/irc-logs</a> / w3c / <a href="irc://irc.w3.org:6665/testing">#testing</a> / 2015-06-29 / <a href="#end">end</a></h1>
<p>Options: 
<p><a href=20150628>Previous day</a>, <a href=20150630>Next day</a></p>
<ol id=lines>
<li id=l-1 class=hide><a href=#l-1>#</a> Session Start: Mon Jun 29 00:00:00 2015</li>
<li id=l-2 class=hide><a href=#l-2>#</a> Session Ident: #testing</li>
<li id=l-3 class=hide><a href=#l-3>#</a> [01:25] * Joins: playjax (~chatzilla@public.cloak)</li>
<li id=l-4 class=hide><a href=#l-4>#</a> [02:50] * Quits: playjax (~chatzilla@public.cloak) (Client closed connection)</li>
<li id=l-5 class=hide><a href=#l-5>#</a> [03:26] * Joins: playjax (~chatzilla@public.cloak)</li>
<li id=l-6 class=hide><a href=#l-6>#</a> [04:29] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-7><a href=#l-7>#</a> [04:29] -gitbot:#testing- [web-platform-tests] sideshowbarker pushed 1 new commit to master: <a href="https://github.com/w3c/web-platform-tests/commit/f1a67fbac028587f9f15e7a1cd141f6c040c8c50">https://github.com/w3c/web-platform-tests/commit/f1a67fbac028587f9f15e7a1cd141f6c040c8c50</a></li>
<li id=l-8><a href=#l-8>#</a> [04:29] -gitbot:#testing- web-platform-tests/master f1a67fb Michael[tm] Smith: [docs] Fixed minor markup error.</li>
<li id=l-9 class=hide><a href=#l-9>#</a> [04:29] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-10 class=hide><a href=#l-10>#</a> [05:24] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-11><a href=#l-11>#</a> [05:24] -gitbot:#testing- [web-platform-tests] sideshowbarker pushed 1 new commit to master: <a href="https://github.com/w3c/web-platform-tests/commit/b90da6aee4c27c41cba6aa4d3e041f3893a3c614">https://github.com/w3c/web-platform-tests/commit/b90da6aee4c27c41cba6aa4d3e041f3893a3c614</a></li>
<li id=l-12><a href=#l-12>#</a> [05:24] -gitbot:#testing- web-platform-tests/master b90da6a Michael[tm] Smith: [cc] Updated some tests w/ invalid CSS media type</li>
<li id=l-13 class=hide><a href=#l-13>#</a> [05:24] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-14 class=hide><a href=#l-14>#</a> [05:52] * Quits: rego (~smuxi@public.cloak) (Ping timeout: 180 seconds)</li>
<li id=l-15 class=hide><a href=#l-15>#</a> [05:54] * Joins: rego (~smuxi@public.cloak)</li>
<li id=l-16 class=hide><a href=#l-16>#</a> [06:20] * Joins: heycam|away (~cam@public.cloak)</li>
<li id=l-17 class=hide _a="heycam|away"><a href=#l-17>#</a> [06:23] * heycam|away is now known as heycam</li>
<li id=l-18 class=hide _a="heycam"><a href=#l-18>#</a> [06:50] * heycam is now known as heycam|away</li>
<li id=l-19 class=hide _a="heycam|away"><a href=#l-19>#</a> [07:42] * heycam|away is now known as heycam</li>
<li id=l-20 class=hide _a="heycam"><a href=#l-20>#</a> [08:41] * heycam is now known as heycam|away</li>
<li id=l-21 class=hide _a="heycam|away"><a href=#l-21>#</a> [09:09] * heycam|away is now known as heycam</li>
<li id=l-22 _a="MikeSmith" _r="jgraham"><a href=#l-22>#</a> [09:17] &lt;MikeSmith&gt; jgraham: <a href="http://www.w3c-test.org/notifications/notification-constructor-invalid.html">http://www.w3c-test.org/notifications/notification-constructor-invalid.html</a> 「Test bug: unrecognized DOMException code &quot;new TypeError()&quot; passed to assert_throws()」</li>
<li id=l-23 _a="MikeSmith"><a href=#l-23>#</a> [09:17] &lt;MikeSmith&gt; wondering what I need to change there</li>
<li id=l-24 _a="MikeSmith"><a href=#l-24>#</a> [09:17] * MikeSmith didn't write that test</li>
<li id=l-25 _a="MikeSmith" _r="oh"><a href=#l-25>#</a> [09:19] &lt;MikeSmith&gt; oh, quotes</li>
<li id=l-26 class=hide><a href=#l-26>#</a> [09:24] * Joins: dom (dom@public.cloak)</li>
<li id=l-27 _a="MikeSmith"><a href=#l-27>#</a> [09:26] &lt;MikeSmith&gt; now I'm wondering why this test times out in Chrome <a href="http://web-platform.test:8000/notifications/event-onclick-manual.html">http://web-platform.test:8000/notifications/event-onclick-manual.html</a></li>
<li id=l-28 _a="MikeSmith"><a href=#l-28>#</a> [09:27] &lt;MikeSmith&gt; <a href="http://w3c-test.org/notifications/event-onclick-manual.html">http://w3c-test.org/notifications/event-onclick-manual.html</a></li>
<li id=l-29 class=hide><a href=#l-29>#</a> [10:19] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-30><a href=#l-30>#</a> [10:19] -gitbot:#testing- [web-platform-tests] sideshowbarker created sideshowbarker/notifications (+1 new commit): <a href="https://github.com/w3c/web-platform-tests/commit/1334634aed3b">https://github.com/w3c/web-platform-tests/commit/1334634aed3b</a></li>
<li id=l-31><a href=#l-31>#</a> [10:19] -gitbot:#testing- web-platform-tests/sideshowbarker/notifications 1334634 Michael[tm] Smith: Fixed a bug in a Web Notifications test.</li>
<li id=l-32 class=hide><a href=#l-32>#</a> [10:19] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-33 class=hide><a href=#l-33>#</a> [10:19] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-34><a href=#l-34>#</a> [10:19] -gitbot:#testing- [web-platform-tests] sideshowbarker opened pull request #1941: Fixed a bug in a Web Notifications test. (master...sideshowbarker/notifications) <a href="https://github.com/w3c/web-platform-tests/pull/1941">https://github.com/w3c/web-platform-tests/pull/1941</a></li>
<li id=l-35 class=hide><a href=#l-35>#</a> [10:19] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-36 _a="MikeSmith" _r="botie"><a href=#l-36>#</a> [10:59] &lt;MikeSmith&gt; botie, inform Ms2ger looking at <a href="http://www.w3c-test.org/media-source/mediasource-seek-during-pending-seek.html">http://www.w3c-test.org/media-source/mediasource-seek-during-pending-seek.html</a> which is causing a harness error in Chrome because it's trying to get a property from something that's undefined. What's the proper testharness.js way to catch something like that and make the test fail? (rather than letting it make the harness error out)</li>
<li id=l-37 _a="botie"><a href=#l-37>#</a> [10:59] &lt;botie&gt; will do</li>
<li id=l-38 class=hide><a href=#l-38>#</a> [11:02] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-39><a href=#l-39>#</a> [11:02] -gitbot:#testing- [web-platform-tests] zqzhang closed pull request #1941: Fixed a bug in a Web Notifications test. (master...sideshowbarker/notifications) <a href="https://github.com/w3c/web-platform-tests/pull/1941">https://github.com/w3c/web-platform-tests/pull/1941</a></li>
<li id=l-40 class=hide><a href=#l-40>#</a> [11:02] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-41 class=hide><a href=#l-41>#</a> [11:02] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-42><a href=#l-42>#</a> [11:02] -gitbot:#testing- [web-platform-tests] zqzhang pushed 1 new commit to master: <a href="https://github.com/w3c/web-platform-tests/commit/5ef62798e92c6b28b0c40d05374d0c1b16ba0718">https://github.com/w3c/web-platform-tests/commit/5ef62798e92c6b28b0c40d05374d0c1b16ba0718</a></li>
<li id=l-43><a href=#l-43>#</a> [11:02] -gitbot:#testing- web-platform-tests/master 5ef6279 Zhiqiang Zhang: Merge pull request #1941 from w3c/sideshowbarker/notifications...</li>
<li id=l-44 class=hide><a href=#l-44>#</a> [11:02] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-45 class=hide><a href=#l-45>#</a> [11:03] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-46><a href=#l-46>#</a> [11:03] -gitbot:#testing- [web-platform-tests] zqzhang deleted sideshowbarker/notifications at 1334634: <a href="https://github.com/w3c/web-platform-tests/commit/1334634">https://github.com/w3c/web-platform-tests/commit/1334634</a></li>
<li id=l-47 class=hide><a href=#l-47>#</a> [11:03] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-48 class=hide _a="heycam"><a href=#l-48>#</a> [11:04] * heycam is now known as heycam|away</li>
<li id=l-49 class=hide><a href=#l-49>#</a> [11:12] * Joins: Ms2ger (~Ms2ger@public.cloak)</li>
<li id=l-50 _a="botie" _r="Ms2ger"><a href=#l-50>#</a> [11:12] &lt;botie&gt; Ms2ger, at 2015-06-29 08:59 UTC, MikeSmith said: looking at <a href="http://www.w3c-test.org/media-source/mediasource-seek-during-pending-seek.html">http://www.w3c-test.org/media-source/mediasource-seek-during-pending-seek.html</a> which is causing a harness error in Chrome because it's trying to get a property from something that's undefined. What's the proper testharness.js way to catch something</li>
<li id=l-51 _a="botie"><a href=#l-51>#</a> [11:12] &lt;botie&gt; like that and make the test fail? (rather than letting it make the harness error out)</li>
<li id=l-52 _a="Ms2ger" _r="MikeSmith"><a href=#l-52>#</a> [11:28] &lt;Ms2ger&gt; MikeSmith, pfff</li>
<li id=l-53 _a="Ms2ger"><a href=#l-53>#</a> [11:28] &lt;Ms2ger&gt; I hate it when people invent their own little harnesses that make the test incomprehensible</li>
<li id=l-54 _a="MikeSmith"><a href=#l-54>#</a> [11:30] &lt;MikeSmith&gt; yeah :(</li>
<li id=l-55 _a="MikeSmith"><a href=#l-55>#</a> [11:30] &lt;MikeSmith&gt; I don't want to look too closely at the rest of that file</li>
<li id=l-56 class=hide><a href=#l-56>#</a> [11:32] * Joins: darobin (rberjon@public.cloak)</li>
<li id=l-57 _a="Ms2ger"><a href=#l-57>#</a> [11:36] &lt;Ms2ger&gt; So yeah, seems like MediaSourceUtil.appendUntilEventFires is busted</li>
<li id=l-58 _a="Ms2ger"><a href=#l-58>#</a> [11:39] * Ms2ger takes an axe to that directory</li>
<li id=l-59 _a="Ms2ger"><a href=#l-59>#</a> [11:42] &lt;Ms2ger&gt; wtf</li>
<li id=l-60 _a="Ms2ger"><a href=#l-60>#</a> [11:42] &lt;Ms2ger&gt; This has its own test runner?</li>
<li id=l-61 class=hide><a href=#l-61>#</a> [11:53] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-62><a href=#l-62>#</a> [11:53] -gitbot:#testing- [web-platform-tests] Ms2ger created media-source (+3 new commits): <a href="https://github.com/w3c/web-platform">https://github.com/w3c/web-platform</a>-tests/compare/21474974b878^...b8ca06754b16</li>
<li id=l-63><a href=#l-63>#</a> [11:53] -gitbot:#testing- web-platform-tests/media-source 2147497 Ms2ger: Remove media-source's test runner....</li>
<li id=l-64><a href=#l-64>#</a> [11:53] -gitbot:#testing- web-platform-tests/media-source 99c1d31 Ms2ger: Disallow tabs in media-source.</li>
<li id=l-65><a href=#l-65>#</a> [11:53] -gitbot:#testing- web-platform-tests/media-source b8ca067 Ms2ger: Disallow trailing whitespace in media-source.</li>
<li id=l-66 class=hide><a href=#l-66>#</a> [11:53] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-67 class=hide><a href=#l-67>#</a> [12:32] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-68><a href=#l-68>#</a> [12:32] -gitbot:#testing- [web-platform-tests] Ms2ger pushed 2 new commits to media-source: <a href="https://github.com/w3c/web-platform-tests/compare/b8ca06754b16...edb207d20806">https://github.com/w3c/web-platform-tests/compare/b8ca06754b16...edb207d20806</a></li>
<li id=l-69><a href=#l-69>#</a> [12:32] -gitbot:#testing- web-platform-tests/media-source 979de36 Ms2ger: Remove media-source's failOnEvent function....</li>
<li id=l-70><a href=#l-70>#</a> [12:32] -gitbot:#testing- web-platform-tests/media-source edb207d Ms2ger: Remove media-source's endOnEvent function....</li>
<li id=l-71 class=hide><a href=#l-71>#</a> [12:32] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-72 class=hide><a href=#l-72>#</a> [12:34] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-73><a href=#l-73>#</a> [12:34] -gitbot:#testing- [web-platform-tests] Ms2ger opened pull request #1942: Cleanup some media source tests. (master...media-source) <a href="https://github.com/w3c/web-platform-tests/pull/1942">https://github.com/w3c/web-platform-tests/pull/1942</a></li>
<li id=l-74 class=hide><a href=#l-74>#</a> [12:34] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-75 _a="Ms2ger" _r="MikeSmith"><a href=#l-75>#</a> [12:34] &lt;Ms2ger&gt; MikeSmith, ^</li>
<li id=l-76 _a="MikeSmith" _r="Ms2ger"><a href=#l-76>#</a> [12:48] &lt;MikeSmith&gt; Ms2ger: on from my mobile atm, will review after a couple hours</li>
<li id=l-77 _a="MikeSmith"><a href=#l-77>#</a> [12:49] &lt;MikeSmith&gt; thanks for doing that</li>
<li id=l-78 _a="MikeSmith"><a href=#l-78>#</a> [12:49] &lt;MikeSmith&gt; those tests need some adult supervision</li>
<li id=l-79 _a="MikeSmith"><a href=#l-79>#</a> [12:50] &lt;MikeSmith&gt; I'm beginning to question the wisdom of our &quot;ok to merge if already reviewed upstream&quot; policy</li>
<li id=l-80 _a="MikeSmith"><a href=#l-80>#</a> [12:51] &lt;MikeSmith&gt; at least in the case of particular browser projects</li>
<li id=l-81 _a="Ms2ger" _r="Still"><a href=#l-81>#</a> [13:10] &lt;Ms2ger&gt; Still, it's better than the alternative</li>
<li id=l-82 class=hide><a href=#l-82>#</a> [13:20] * Joins: ArtB (~ArtB@public.cloak)</li>
<li id=l-83 _a="burnik"><a href=#l-83>#</a> [13:58] * burnik slaps jgraham around a bit with a large fishbot</li>
<li id=l-84 class=hide><a href=#l-84>#</a> [14:15] * Joins: zcorpan (~zcorpan@public.cloak)</li>
<li id=l-85 class=hide><a href=#l-85>#</a> [14:21] * Quits: zcorpan (~zcorpan@public.cloak) (Client closed connection)</li>
<li id=l-86 class=hide><a href=#l-86>#</a> [14:21] * Joins: zcorpan (~zcorpan@public.cloak)</li>
<li id=l-87 class=hide><a href=#l-87>#</a> [14:24] * Quits: zcorpan (~zcorpan@public.cloak) (Client closed connection)</li>
<li id=l-88 class=hide><a href=#l-88>#</a> [14:39] * Joins: zcorpan (~zcorpan@public.cloak)</li>
<li id=l-89 class=hide><a href=#l-89>#</a> [14:55] * Quits: zcorpan (~zcorpan@public.cloak) (Client closed connection)</li>
<li id=l-90 class=hide><a href=#l-90>#</a> [15:40] * Quits: playjax (~chatzilla@public.cloak) (Client closed connection)</li>
<li id=l-91 class=hide><a href=#l-91>#</a> [15:42] * Joins: plh (plehegar@public.cloak)</li>
<li id=l-92 _a="MikeSmith"><a href=#l-92>#</a> [15:56] &lt;MikeSmith&gt; hmm our github IRC bot not working as expected atm</li>
<li id=l-93 _a="darobin" _r="MikeSmith"><a href=#l-93>#</a> [15:56] &lt;darobin&gt; MikeSmith: which one, gitbot?</li>
<li id=l-94 _a="MikeSmith" _r="darobin"><a href=#l-94>#</a> [15:57] &lt;MikeSmith&gt; darobin: yeah <a href="https://github.com/w3c/web-platform-tests/settings/hooks/721165">https://github.com/w3c/web-platform-tests/settings/hooks/721165</a></li>
<li id=l-95 _a="MikeSmith"><a href=#l-95>#</a> [15:57] &lt;MikeSmith&gt; admin UI is saying &quot;Last delivery was not successful. Service Timeout.&quot;</li>
<li id=l-96 _a="darobin" _r="MikeSmith"><a href=#l-96>#</a> [15:59] &lt;darobin&gt; MikeSmith: I don't think that's anything we can do anything about; it's all GitHub</li>
<li id=l-97 _a="darobin"><a href=#l-97>#</a> [16:00] &lt;darobin&gt; maybe there was network trouble, zcorpan sure seems to have been on/off</li>
<li id=l-98 _a="MikeSmith"><a href=#l-98>#</a> [16:01] &lt;MikeSmith&gt; hope w3c isn't blocking it somehow</li>
<li id=l-99 _a="MikeSmith"><a href=#l-99>#</a> [16:01] &lt;MikeSmith&gt; it was working still ~3.5 hours ago</li>
<li id=l-100 _a="MikeSmith"><a href=#l-100>#</a> [16:03] &lt;MikeSmith&gt; anyway Ms2ger I merged that PR</li>
<li id=l-101 _a="MikeSmith"><a href=#l-101>#</a> [16:03] &lt;MikeSmith&gt; thanks again</li>
<li id=l-102 _a="MikeSmith"><a href=#l-102>#</a> [16:04] &lt;MikeSmith&gt; but now we can get back to the part where you help me understand how to fix that one test so it doesn't cause the harness to error out</li>
<li id=l-103 class=hide><a href=#l-103>#</a> [17:02] * Quits: dom (dom@public.cloak) (Client closed connection)</li>
<li id=l-104 _a="Ms2ger" _r="Yeah"><a href=#l-104>#</a> [17:50] &lt;Ms2ger&gt; Yeah, I have no idea</li>
<li id=l-105 class=hide><a href=#l-105>#</a> [18:30] * Quits: Ms2ger (~Ms2ger@public.cloak) (Client closed connection)</li>
<li id=l-106 class=hide><a href=#l-106>#</a> [18:44] * Joins: bhill2 (~bhill2@public.cloak)</li>
<li id=l-107 class=hide><a href=#l-107>#</a> [18:46] * Joins: Ms2ger (~Ms2ger@public.cloak)</li>
<li id=l-108 class=hide><a href=#l-108>#</a> [18:50] * Joins: dom (dom@public.cloak)</li>
<li id=l-109 class=hide><a href=#l-109>#</a> [18:58] * Quits: bhill2 (~bhill2@public.cloak) (Client closed connection)</li>
<li id=l-110 class=hide><a href=#l-110>#</a> [19:02] * Joins: bhill2 (~bhill2@public.cloak)</li>
<li id=l-111 class=hide><a href=#l-111>#</a> [19:03] * Quits: rego (~smuxi@public.cloak) (Ping timeout: 180 seconds)</li>
<li id=l-112 class=hide><a href=#l-112>#</a> [19:04] * Quits: bhill2 (~bhill2@public.cloak) (Client closed connection)</li>
<li id=l-113 class=hide><a href=#l-113>#</a> [19:04] * Joins: bhill2 (~bhill2@public.cloak)</li>
<li id=l-114 class=hide><a href=#l-114>#</a> [19:05] * Joins: rego (~smuxi@public.cloak)</li>
<li id=l-115 class=hide><a href=#l-115>#</a> [19:11] * Quits: darobin (rberjon@public.cloak) (Client closed connection)</li>
<li id=l-116 class=hide _a="terri"><a href=#l-116>#</a> [19:25] * terri is now known as terri_offline</li>
<li id=l-117 class=hide><a href=#l-117>#</a> [19:29] * Quits: dom (dom@public.cloak) (&quot;&quot;)</li>
<li id=l-118 class=hide><a href=#l-118>#</a> [19:32] * Quits: bhill2 (~bhill2@public.cloak) (Client closed connection)</li>
<li id=l-119 class=hide _a="terri_offline"><a href=#l-119>#</a> [19:45] * terri_offline is now known as terri</li>
<li id=l-120 class=hide><a href=#l-120>#</a> [20:13] * Joins: hober (~ted@public.cloak)</li>
<li id=l-121 class=hide><a href=#l-121>#</a> [20:18] * Joins: bhill2 (~bhill2@public.cloak)</li>
<li id=l-122 class=hide><a href=#l-122>#</a> [20:21] * Joins: darobin (rberjon@public.cloak)</li>
<li id=l-123 class=hide><a href=#l-123>#</a> [20:26] * Quits: bhill2 (~bhill2@public.cloak) (Client closed connection)</li>
<li id=l-124 class=hide><a href=#l-124>#</a> [20:27] * Quits: darobin (rberjon@public.cloak) (Client closed connection)</li>
<li id=l-125 class=hide><a href=#l-125>#</a> [20:37] * Joins: bhill2 (~bhill2@public.cloak)</li>
<li id=l-126 class=hide><a href=#l-126>#</a> [20:44] * Quits: bhill2 (~bhill2@public.cloak) (Ping timeout: 180 seconds)</li>
<li id=l-127 class=hide><a href=#l-127>#</a> [20:51] * Joins: darobin (rberjon@public.cloak)</li>
<li id=l-128 class=hide><a href=#l-128>#</a> [21:50] * Joins: bhill2 (~bhill2@public.cloak)</li>
<li id=l-129 class=hide><a href=#l-129>#</a> [22:01] * Quits: bhill2 (~bhill2@public.cloak) (Client closed connection)</li>
<li id=l-130 class=hide><a href=#l-130>#</a> [22:02] * Joins: bhill2 (~bhill2@public.cloak)</li>
<li id=l-131 class=hide><a href=#l-131>#</a> [22:03] * Quits: bhill2 (~bhill2@public.cloak) (Client closed connection)</li>
<li id=l-132 class=hide><a href=#l-132>#</a> [22:05] * Joins: bhill2 (~bhill2@public.cloak)</li>
<li id=l-133 class=hide><a href=#l-133>#</a> [22:12] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-134><a href=#l-134>#</a> [22:12] -gitbot:#testing- [web-platform-tests] jgraham created sync_b6ac93b601e39fd004d484c0aefbfda01187008b (+1 new commit): <a href="https://github.com/w3c/web-platform-tests/commit/b6ac93b601e3">https://github.com/w3c/web-platform-tests/commit/b6ac93b601e3</a></li>
<li id=l-135><a href=#l-135>#</a> [22:12] -gitbot:#testing- web-platform-tests/sync_b6ac93b601e39fd004d484c0aefbfda01187008b b6ac93b Karl Tomlinson: adjust explict revoke test to fail if revocation fails...</li>
<li id=l-136 class=hide><a href=#l-136>#</a> [22:12] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-137 class=hide><a href=#l-137>#</a> [22:12] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-138><a href=#l-138>#</a> [22:12] -gitbot:#testing- [web-platform-tests] jgraham closed pull request #1943: adjust explict revoke test to fail if revocation fails (master...sync_b6ac93b601e39fd004d484c0aefbfda01187008b) <a href="https://github.com/w3c/web-platform-tests/pull/1943">https://github.com/w3c/web-platform-tests/pull/1943</a></li>
<li id=l-139 class=hide><a href=#l-139>#</a> [22:12] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-140 class=hide><a href=#l-140>#</a> [22:12] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-141><a href=#l-141>#</a> [22:12] -gitbot:#testing- [web-platform-tests] jgraham deleted sync_b6ac93b601e39fd004d484c0aefbfda01187008b at b6ac93b: <a href="https://github.com/w3c/web-platform-tests/commit/b6ac93b">https://github.com/w3c/web-platform-tests/commit/b6ac93b</a></li>
<li id=l-142 class=hide><a href=#l-142>#</a> [22:12] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-143 class=hide><a href=#l-143>#</a> [22:12] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-144><a href=#l-144>#</a> [22:12] -gitbot:#testing- [web-platform-tests] jgraham opened pull request #1944: test opening referenced MediaSource after URL.revokeObjectURL() (master...sync_c74cc59d094cca6302700f9e99be6185389923b6) <a href="https://github.com/w3c/web-platform-tests/pull/1944">https://github.com/w3c/web-platform-tests/pull/1944</a></li>
<li id=l-145 class=hide><a href=#l-145>#</a> [22:12] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-146 class=hide><a href=#l-146>#</a> [22:12] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-147><a href=#l-147>#</a> [22:12] -gitbot:#testing- [web-platform-tests] jgraham deleted sync_c74cc59d094cca6302700f9e99be6185389923b6 at c74cc59: <a href="https://github.com/w3c/web-platform-tests/commit/c74cc59">https://github.com/w3c/web-platform-tests/commit/c74cc59</a></li>
<li id=l-148 class=hide><a href=#l-148>#</a> [22:12] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-149 class=hide><a href=#l-149>#</a> [22:12] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-150><a href=#l-150>#</a> [22:12] -gitbot:#testing- [web-platform-tests] jgraham opened pull request #1945: test auto-revoking behavior with URL.createObjectURL(MediaSource) (master...sync_7ca2c25204985477514ddb4e9171e7544d98a8e5) <a href="https://github.com/w3c/web-platform-tests/pull/1945">https://github.com/w3c/web-platform-tests/pull/1945</a></li>
<li id=l-151 class=hide><a href=#l-151>#</a> [22:12] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-152 class=hide><a href=#l-152>#</a> [22:12] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-153><a href=#l-153>#</a> [22:12] -gitbot:#testing- [web-platform-tests] jgraham closed pull request #1945: test auto-revoking behavior with URL.createObjectURL(MediaSource) (master...sync_7ca2c25204985477514ddb4e9171e7544d98a8e5) <a href="https://github.com/w3c/web-platform-tests/pull/1945">https://github.com/w3c/web-platform-tests/pull/1945</a></li>
<li id=l-154 class=hide><a href=#l-154>#</a> [22:12] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-155 class=hide><a href=#l-155>#</a> [22:12] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-156><a href=#l-156>#</a> [22:12] -gitbot:#testing- [web-platform-tests] jgraham created sync_1347ee9544bb40827e887a57f479c6affe132293 (+1 new commit): <a href="https://github.com/w3c/web-platform-tests/commit/1347ee9544bb">https://github.com/w3c/web-platform-tests/commit/1347ee9544bb</a></li>
<li id=l-157><a href=#l-157>#</a> [22:12] -gitbot:#testing- web-platform-tests/sync_1347ee9544bb40827e887a57f479c6affe132293 1347ee9 James Graham: Rename gainnode test to work around treeherder limitation,...</li>
<li id=l-158 class=hide><a href=#l-158>#</a> [22:12] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-159 class=hide><a href=#l-159>#</a> [22:12] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-160><a href=#l-160>#</a> [22:12] -gitbot:#testing- [web-platform-tests] jgraham closed pull request #1946: Rename gainnode test to work around treeherder limitation, (master...sync_1347ee9544bb40827e887a57f479c6affe132293) <a href="https://github.com/w3c/web-platform-tests/pull/1946">https://github.com/w3c/web-platform-tests/pull/1946</a></li>
<li id=l-161 class=hide><a href=#l-161>#</a> [22:12] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-162 class=hide><a href=#l-162>#</a> [22:13] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-163><a href=#l-163>#</a> [22:13] -gitbot:#testing- [web-platform-tests] jgraham created sync_0a5be2de9ea51afc9df3802679972d41619ee619 (+1 new commit): <a href="https://github.com/w3c/web-platform-tests/commit/0a5be2de9ea5">https://github.com/w3c/web-platform-tests/commit/0a5be2de9ea5</a></li>
<li id=l-164><a href=#l-164>#</a> [22:13] -gitbot:#testing- web-platform-tests/sync_0a5be2de9ea51afc9df3802679972d41619ee619 0a5be2d Ben Turner: Fix IndexedDB tests to expect InvalidStateError rather than TransactionInactiveError when touching a deleted IDBObjectStore or IDBIndex object, r=janv....</li>
<li id=l-165 class=hide><a href=#l-165>#</a> [22:13] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-166 class=hide><a href=#l-166>#</a> [22:13] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-167><a href=#l-167>#</a> [22:13] -gitbot:#testing- [web-platform-tests] jgraham closed pull request #1947: Fix IndexedDB tests to expect InvalidStateError rather than TransactionInactiveError when touching a deleted IDBObjectStore or IDBIndex object, r=janv. (master...sync_0a5be2de9ea51afc9df3802679972d41619ee619) <a href="https://github.com/w3c/web-platform-tests/pull/1947">https://github.com/w3c/web-platform-tests/pull/1947</a></li>
<li id=l-168 class=hide><a href=#l-168>#</a> [22:13] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-169 class=hide><a href=#l-169>#</a> [22:13] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-170><a href=#l-170>#</a> [22:13] -gitbot:#testing- [web-platform-tests] jgraham deleted sync_0a5be2de9ea51afc9df3802679972d41619ee619 at 0a5be2d: <a href="https://github.com/w3c/web-platform-tests/commit/0a5be2d">https://github.com/w3c/web-platform-tests/commit/0a5be2d</a></li>
<li id=l-171 class=hide><a href=#l-171>#</a> [22:13] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-172 class=hide><a href=#l-172>#</a> [22:13] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-173><a href=#l-173>#</a> [22:13] -gitbot:#testing- [web-platform-tests] jgraham created sync_0126c2f5efd77dca1530305295fd27db15a9654f (+1 new commit): <a href="https://github.com/w3c/web-platform-tests/commit/0126c2f5efd7">https://github.com/w3c/web-platform-tests/commit/0126c2f5efd7</a></li>
<li id=l-174><a href=#l-174>#</a> [22:13] -gitbot:#testing- web-platform-tests/sync_0126c2f5efd77dca1530305295fd27db15a9654f 0126c2f Andrea Marchesini: patch 2 - Make searchParams attribute readonly,...</li>
<li id=l-175 class=hide><a href=#l-175>#</a> [22:13] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-176 class=hide><a href=#l-176>#</a> [22:13] * Joins: gitbot (~gitbot@public.cloak)</li>
<li id=l-177><a href=#l-177>#</a> [22:13] -gitbot:#testing- [web-platform-tests] jgraham closed pull request #1948: patch 2 - Make searchParams attribute readonly, (master...sync_0126c2f5efd77dca1530305295fd27db15a9654f) <a href="https://github.com/w3c/web-platform-tests/pull/1948">https://github.com/w3c/web-platform-tests/pull/1948</a></li>
<li id=l-178 class=hide><a href=#l-178>#</a> [22:13] * Parts: gitbot (~gitbot@public.cloak)</li>
<li id=l-179 class=hide><a href=#l-179>#</a> [22:25] * Quits: darobin (rberjon@public.cloak) (Client closed connection)</li>
<li id=l-180 class=hide><a href=#l-180>#</a> [22:59] * Quits: bhill2 (~bhill2@public.cloak) (Client closed connection)</li>
<li id=l-181 class=hide><a href=#l-181>#</a> [23:06] * Joins: bhill2 (~bhill2@public.cloak)</li>
<li id=l-182 class=hide><a href=#l-182>#</a> [23:53] * Joins: darobin (rberjon@public.cloak)</li>
<li id=l-183 class=hide><a href=#l-183>#</a> [23:57] * Quits: darobin (rberjon@public.cloak) (Client closed connection)</li>
<li id=l-184 class=hide><a href=#l-184>#</a> [23:59] * Joins: darobin (rberjon@public.cloak)</li>
<li id=l-185 class=hide><a href=#l-185>#</a> Session Close: Tue Jun 30 00:00:00 2015</li>
</ol>
<p id=end><a href=20150628>Previous day</a>, <a href=20150630>Next day</a></p>
<p>Think <a href="/irc-logs/">these logs</a> are useful? Then <a href="https://www.paypal.com/cgi-bin/webscr?cmd=_donations&amp;business=PMXKLAEH53JSS&amp;lc=GB&amp;item_name=Krijn%20Hoetmer&amp;item_number=IRC%20Logs&amp;currency_code=EUR&amp;bn=PP%2dDonationsBF%3abtn_donate_SM%2egif%3aNonHosted"><strong>please donate</strong></a> to show your gratitude (and keep them up, of course). <a href="https://twitter.com/krijnhoetmer/status/414824591337734144">Thanks! — Krijn</a></p>
<script src=/irc-logs/s2.js></script>
